pyproject.toml
[tool.poetry]
name = "auth_service"
version = "0.1.0"
description = "Auth_service — Микросервис для аутентификации, авторизации, управления токенами."
authors = ["S.Ryzhakov <sergo.ryzakov@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
sqlalchemy = ">=2.0.41,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.16.4,<2.0.0"
redis = ">=6.2.0,<7.0.0"
aio-pika = ">=9.5.5,<10.0.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = ">=2.10.1,<3.0.0"
dependency-injector = ">=4.48.1,<5.0.0"
polyfactory = "^2.22.1"
colorama = "^0.4.6"
httpx = "^0.28.1"
psycopg2-binary = "^2.9.10"
autoflake = "^2.3.1"
docformatter = "^1.7.7"
add-trailing-comma = "^3.2.0"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "<4.1.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true
known_first_party = ["app"]
include_trailing_comma = true

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
ignore-init-module-imports = true
in-place = true
recursive = true

[tool.docformatter]
wrap-summaries = 100
wrap-descriptions = 100
pre-summary-newline = true
make-summary-multi-line = true
blank = true
